pyproject.toml
[tool.poetry]
name = "mh-clickhouse-driver"
version = "0.3.0"
description = ""
authors = ["Vladimir Trifonov <3fonov@gmail.com>"]
readme = "README.md"
packages = [{include = "mh_clickhouse_driver"}]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.5.2"
requests = "^2.28.1"
clickhouse-driver = "^0.2.5"
numpy = "^1.23.5"
python-dotenv = "^0.21.0"
google-cloud-bigquery = "^3.4.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
autopep8 = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
